Cards can be added to the scene without any problem. However, newing them does not work as if they can not be added to the scene

+++ cpp files/Tower.cpp
@@ -4,7 +4,7 @@
 
 #include <header files/Tower.h>
 
-Tower::Tower(int tp, int trgt, double hs, int hp, int d, double r, int tr, QPixmap *pic, QTimer *timer) :
+Tower::Tower(int tp, int trgt, double hs, int hp, int d, double r, double tr, QPixmap *pic, QTimer *timer) :
             Object(tp, trgt, hs, hp, d, r, tr, timer)
 {
     this->setPixmap(*pic);
